fit_data_loss | fix checkstyle warnings
